translator/requirements.txt
python-dotenv
assemblyai
translate
elevenlabs